+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## [1.1.0](https://github.com/nuxt-community/partytown-module/compare/1.0.2...1.1.0) (2022-02-04)
+
+
+### Features
+
+* normalize lib option ([b2875f3](https://github.com/nuxt-community/partytown-module/commit/b2875f3c511866d6f355f611797178c2e4cdaaf3))
+* upgrade to [@builder](https://github.com/builder).io/partytown v0.3.1 ([f2a05aa](https://github.com/nuxt-community/partytown-module/commit/f2a05aaf8ce4f332416cc5a7891e189763956380))
+
+
+### Bug Fixes
+
+* inherit partytown configuration types ([f28ee13](https://github.com/nuxt-community/partytown-module/commit/f28ee13a009a515e87ad17700b07388de765a247))
+
 ### [1.0.2](https://github.com/nuxt-community/partytown-module/compare/1.0.1...1.0.2) (2022-02-03)
 
 ### 1.0.1 (2022-02-03)